+++ crates/test_design/src/test/validate/mod.rs
@@ -1,8 +1,10 @@
 pub mod case;
 pub mod collection;
 pub mod decimal;
+pub mod finance;
 pub mod list;
 pub mod newtype;
+pub mod option;
 pub mod record;
 
 use crate::prelude::*;